feat(generator) save images endpoint

+++ src/app/use-cases/generate-image-of-service.use-case.ts
@@ -13,15 +13,55 @@ export class GenerateImageOfServiceUseCase {
 
   async execute(data: GenerateImageOfServiceDto): Promise<[string, string]> {
     const { numberOfImages, companyName, serviceName } = data;
-    const prompt = buildPromptFromTemplate(data);
-    const imagePath = await this.generatorService.generateImage(prompt, numberOfImages);
-    let imageName = imagePath.split('/').pop();
-    imageName = imageName?.split('\\').pop();
-    const uploadPath = await this.storageService.upload({
-      fileName: companyName + '/' + serviceName + '/' + imageName,
-      filePath: imagePath,
+    // verify if the companyName and serviceName are not empty
+    if (!companyName || !serviceName) {
+      throw new Error('Company name and service name are required');
+    }
+    // verify if exist image in storage and if it is used
+    const [existImage, images] = await this.existImagesUnused(companyName, serviceName);
+    if (!existImage) {
+      // if it is used, generate a new image
+      const prompt = buildPromptFromTemplate(data);
+      const imagePath = await this.generatorService.generateImage(prompt, numberOfImages);
+      let imageName = imagePath.split('/').pop();
+      imageName = imageName?.split('\\').pop();
+      const uploadPath = await this.storageService.upload({
+        fileName: companyName + '/' + serviceName + '/' + imageName,
+        filePath: imagePath,
+        rootFolder: 'IA_IMAGES/',
+      });
+      // rename the file in storage to mark it as used
+      await this.storageService.renameFile({
+        fileName: companyName + '/' + serviceName + '/' + imageName,
+        newFileName: companyName + '/' + serviceName + '/' + imageName?.replace('.jpg', '_used.jpg'),
+        rootFolder: 'IA_IMAGES/',
+      });
+      // delete the image from the generator service
+      await this.generatorService.deleteImage(imagePath);
+      return [imagePath, uploadPath];
+    } else {
+      // if it is not used, download the image
+      const imageName = images[0].split('/').pop();
+      const uploadPath = await this.storageService.download({
+        fileName: companyName + '/' + serviceName + '/' + imageName,
+        rootFolder: 'IA_IMAGES/',
+      });
+      // rename the file in storage to mark it as used
+      await this.storageService.renameFile({
+        fileName: companyName + '/' + serviceName + '/' + imageName,
+        newFileName: companyName + '/' + serviceName + '/' + imageName?.replace('.jpg', '_used.jpg'),
+        rootFolder: 'IA_IMAGES/',
+      });
+      return [images[0], uploadPath];
+    }
+  }
+
+  async existImagesUnused(companyName: string, serviceName: string): Promise<[boolean, string[]]> {
+    const images = await this.storageService.getImages({
+      prefix: `${companyName}/${serviceName}/`,
+      rootFolder: 'IA_IMAGES/',
     });
-    await this.generatorService.deleteImage(imagePath);
-    return [imagePath, uploadPath];
+    const unusedImages = images.filter((image) => !image.includes('_used.jpg'));
+    return [unusedImages.length > 0, unusedImages];
   }
 }
